dagster-poc/ingestion/pyproject.toml
[project]
name = "ingestion"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster>=1.12.8",
    "dagster-databricks>=0.28.0",
    "dagster-sling>=0.28.0",
    "dagster-dlt>=0.28.1",
    "dlt[postgres]>=1.5.0",
    "dlt[databricks]>=1.5.0",
    "pandas>=2.3.3,<4",
    "pyarrow>=14.0.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "ruff",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingestion"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "ingestion"
registry_modules = [
    "ingestion.components.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
